+++ Foundation/SpecHelper/Fakes/PSHKFakeOperationQueue.h
@@ -1,0 +1,10 @@
+#import <Foundation/Foundation.h>
+
+@interface PSHKFakeOperationQueue : NSOperationQueue
+
+@property (nonatomic) BOOL runSynchronously;
+
+- (void)reset;
+- (void)runNextOperation;
+
+@end