+++ Foundation/SpecHelper/Foundation+PivotalSpecHelper.h
@@ -3,4 +3,4 @@
 #import "PSHKFakeHTTPURLResponse.h"
 #import "PSHKFakeResponses.h"
 #import "PSHKFixtures.h"
-#import "FakeOperationQueue.h"
+#import "PSHKFakeOperationQueue.h"